Move Timer natures from edge.core to edge.common
