+++ io.openems.edge.common/src/io/openems/edge/common/timer/AbstractTimer.java
@@ -1,9 +1,8 @@
-package io.openems.edge.core.timer;
+package io.openems.edge.common.timer;
 
 import java.util.Optional;
 
 import io.openems.edge.common.channel.Channel;
-import io.openems.edge.common.timer.Timer;
 
 public abstract class AbstractTimer implements Timer {
