+++ io.openems.edge.common/src/io/openems/edge/common/timer/TimerByCount.java
@@ -1,7 +1,6 @@
-package io.openems.edge.core.timer;
+package io.openems.edge.common.timer;
 
 import io.openems.edge.common.channel.Channel;
-import io.openems.edge.common.timer.Timer;
 
 public class TimerByCount extends AbstractTimer implements Timer {
