+++ io.openems.edge.controller.evcs.vehiclebatterysymmetric/src/io/openems/edge/controller/evcs/vehiclebatterysymmetric/ControllerManagedVehicleBatterySymmetricImpl.java
@@ -18,8 +18,8 @@ import io.openems.edge.common.component.AbstractOpenemsComponent;
 import io.openems.edge.common.component.OpenemsComponent;
 import io.openems.edge.common.sum.Sum;
 import io.openems.edge.common.timer.Timer;
+import io.openems.edge.common.timer.TimerManager;
 import io.openems.edge.controller.api.Controller;
-import io.openems.edge.core.timer.TimerManager;
 import io.openems.edge.ess.api.SymmetricEss;
 import io.openems.edge.evcs.api.ManagedVehicleBattery;
 import io.openems.edge.evcs.api.Status;
