--- io.openems.edge.core/src/io/openems/edge/core/timer/package-info.java
@@ -1,3 +1,0 @@
-@org.osgi.annotation.versioning.Version("1.0.0")
-@org.osgi.annotation.bundle.Export
-package io.openems.edge.core.timer;
