+++ io.openems.edge.core/test/io/openems/edge/core/appmanager/Apps.java
@@ -29,11 +29,11 @@ import io.openems.edge.app.heat.HeatingElement;
 import io.openems.edge.app.integratedsystem.FeneconHome;
 import io.openems.edge.app.integratedsystem.FeneconHome20;
 import io.openems.edge.app.integratedsystem.FeneconHome30;
-import io.openems.edge.app.loadcontrol.ManualRelayControl;
-import io.openems.edge.app.loadcontrol.ThresholdControl;
 import io.openems.edge.app.integratedsystem.fenecon.industrial.s.Isk010;
 import io.openems.edge.app.integratedsystem.fenecon.industrial.s.Isk011;
 import io.openems.edge.app.integratedsystem.fenecon.industrial.s.Isk110;
+import io.openems.edge.app.loadcontrol.ManualRelayControl;
+import io.openems.edge.app.loadcontrol.ThresholdControl;
 import io.openems.edge.app.meter.MicrocareSdm630Meter;
 import io.openems.edge.app.meter.SocomecMeter;
 import io.openems.edge.app.peakshaving.PeakShaving;
