+++ io.openems.edge.evcs.cluster.chargemanagement/src/io/openems/edge/evcs/cluster/chargemanagement/EvcsClusterChargeMgmtImpl.java
@@ -26,8 +26,8 @@ import io.openems.edge.common.channel.BooleanWriteChannel;
 import io.openems.edge.common.component.ComponentManager;
 import io.openems.edge.common.component.OpenemsComponent;
 import io.openems.edge.common.event.EdgeEventConstants;
+import io.openems.edge.common.timer.TimerManager;
 import io.openems.edge.controller.evcs.cluster.chargemanagement.EvcsClusterLimiter;
-import io.openems.edge.core.timer.TimerManager;
 import io.openems.edge.evcs.api.Evcs;
 import io.openems.edge.evcs.api.EvcsPower;
 import io.openems.edge.evcs.api.ManagedEvcs;
