+++ io.openems.edge.evcs.compleo/src/io/openems/edge/evcs/compleo/EvcsCompleoImpl.java
@@ -41,8 +41,8 @@ import io.openems.edge.common.component.OpenemsComponent;
 import io.openems.edge.common.event.EdgeEventConstants;
 import io.openems.edge.common.taskmanager.Priority;
 import io.openems.edge.common.timer.Timer;
+import io.openems.edge.common.timer.TimerManager;
 import io.openems.edge.common.type.TypeUtils;
-import io.openems.edge.core.timer.TimerManager;
 import io.openems.edge.evcs.api.ChargeStateHandler;
 import io.openems.edge.evcs.api.ChargingType;
 import io.openems.edge.evcs.api.Evcs;
