+++ io.openems.edge.evcs.compleo/test/io/openems/edge/evcs/compleo/EvcsCompleoImplTest.java
@@ -9,11 +9,11 @@ import io.openems.common.test.TimeLeapClock;
 import io.openems.common.types.ChannelAddress;
 import io.openems.edge.bridge.modbus.test.DummyModbusBridge;
 import io.openems.edge.common.test.AbstractComponentTest.TestCase;
+import io.openems.edge.common.timer.DummyTimerManager;
+import io.openems.edge.common.timer.TimerManager;
 import io.openems.edge.common.test.ComponentTest;
 import io.openems.edge.common.test.DummyComponentManager;
 import io.openems.edge.common.test.DummyConfigurationAdmin;
-import io.openems.edge.core.timer.DummyTimerManager;
-import io.openems.edge.core.timer.TimerManager;
 import io.openems.edge.evcs.api.PhaseRotation;
 import io.openems.edge.evcs.api.Priority;
 import io.openems.edge.evcs.api.Status;
